+++ cmsplugin_blocks/__init__.py
@@ -1,2 +1,2 @@
 """Some basic component plugins to make content with less involved HTML than directly using CKeditor for everything."""
-__version__ = '0.4.3'
+__version__ = '0.5.0-pre.1'