+++ cmsplugin_blocks/forms/album.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-import zipfile
-
 from django.conf import settings
 from django import forms
 from django.utils.translation import ugettext_lazy as _
@@ -9,7 +7,8 @@ from django.forms.widgets import FileInput, NumberInput
 from djangocms_text_ckeditor.widgets import TextEditorWidget
 
 from cmsplugin_blocks.models.album import Album, AlbumItem
-from cmsplugin_blocks.utils import store_images_from_zip
+from cmsplugin_blocks.utils import (validate_file_size, validate_zip,
+                                    store_images_from_zip)
 
 
 class AlbumItemForm(forms.ModelForm):
@@ -44,26 +43,14 @@ class AlbumForm(forms.ModelForm):
 
     def clean_mass_upload(self):
         """
-        Validate uploaded ZIP archive file and temporary store it in
-        'uploaded_zip' form object attribute if valid
+        Validate uploaded ZIP archive file and temporary store it to
+        'uploaded_zip' form object attribute if valid.
         """
         data = self.cleaned_data['mass_upload']
 
         if data:
-            # Validate ZIP from file metas first octets
-            if not zipfile.is_zipfile(data):
-                raise forms.ValidationError("Submited file is not a ZIP archive file")
-
-            # Open ZIP
-            archive = zipfile.ZipFile(data)
-
-            # Search for corrupted files
-            corrupted_file = archive.testzip()
-            if corrupted_file:
-                raise forms.ValidationError("File '{}' in ZIP archive is corrupted".format(corrupted_file))
-
-            # ZIP is totally ok, store it to attribute
-            self.uploaded_zip = archive
+            validate_file_size(data)
+            validate_zip(data, obj=self)
 
         return data